# Do not cache instances of `Compiler` in `CompilerStack`