~Depends on #15451.~ Merged.
Related to #15459.

Another follow-up to #15451, this time removing the `Compiler` instances stored in `CompilerStack`.

This is a test PR to benchmark it see if it's even worth it. But TBH storing these objects in CompilerStack seems like a weird design choice. Apparently the only thing we need from them are the generated sources. And it's not even like they're generated on demand - they're stored in compiler context and simply copied from there. For this reason I'd be in favor of merging this even it does not improve performance. Just not making it worse would be enough.