@cameel 2024-09-26
### Benchmarks
Results from `external.sh`. #15451 vs this PR.

|         File | Pipeline | Time (before) | Time (after) | Memory (after) | Memory (after) |
|--------------|----------|--------------:|-------------:|---------------:|---------------:|
| openzeppelin |   legacy |         10  s |        10  s |       517  MiB |      514  MiB  |
| openzeppelin |       ir |         33  s |        33  s |       497  MiB |      495  MiB  |
| uniswap-v4   |   legacy |         21  s |        20  s |       1012 MiB |      1007 MiB  |
| uniswap-v4   |       ir |         119 s |        120 s |       1494 MiB |      1499 MiB  |
| eigenlayer   |   legacy |         74  s |        77  s |       2835 MiB |      2821 MiB  |
| eigenlayer   |       ir |         522 s |        517 s |       4457 MiB |      4471 MiB  |

Looks like it does not make much difference, both in terms of memory and execution time.