@cameel 2024-09-26
I added a commit that removes the `generatedSources` and `runtimeGeneratedSources` fields, storing JSON version of generated sources in a `LazyInit` container (copied over from one #15459).

With this results seem on par with https://github.com/ethereum/solidity/pull/15459#issuecomment-2377932788.

### Benchmarks after `LazyInit` removal
Results from `external.sh`, 1 run. #15451 vs this PR.

#### Legacy
|         File | Time (before) | Time (after) | Memory (before) | Memory (after) |
|--------------|--------------:|-------------:|----------------:|---------------:|
| openzeppelin |           9 s |         11 s |         517 MiB |        454 MiB |
| uniswap-v4   |          20 s |         20 s |        1012 MiB |        902 MiB |
| eigenlayer   |          72 s |         76 s |        2833 MiB |       2411 MiB |
| seaport      |         150 s |        150 s |        4477 MiB |       3896 MiB |
| sablier-v2   |         141 s |        141 s |        5215 MiB |       5020 MiB |

#### IR
|         File | Time (before) | Time (after) | Memory (before) | Memory (after) |
|--------------|--------------:|-------------:|----------------:|---------------:|
| openzeppelin |          31 s |         35 s |         497 MiB |        496 MiB |
| uniswap-v4   |         117 s |        127 s |        1491 MiB |       1498 MiB |
| eigenlayer   |         509 s |        515 s |        4455 MiB |       4461 MiB |

The running time here seemingly increased, but I suspect it's just a fluke. Will rerun to make sure.